uffd: uffd_open prints info, caller prints error
When uffd_open is called from kerndat_uffd, userfaultfd failiure is not
considered an error, so the goal is to suppress the error message --
instead, the goal of this change is to print this message as info.

Initially, uffd_open handles userfaultfd fails with errors only, but this
implementation distinguishes whether to report as an error or info
depending on where uffd_open was called.

Signed-off-by: Angie Ni <avtni@google.com>